General code clean up - moved dto to domain/dto - renamed ProductGroupDto to GroupDto - general code clean up
