+++ src/main/java/com/kodilla/ecommercee/config/CoreConfiguration.java
@@ -1,5 +1,4 @@
 package com.kodilla.ecommercee.config;
-
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
