+++ src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -1,12 +1,13 @@
 package com.kodilla.ecommercee.controller;
-import com.kodilla.ecommercee.dto.UserDto;
+
+import com.kodilla.ecommercee.domain.dto.UserDto;
 import com.kodilla.ecommercee.mapper.UserMapper;
 import com.kodilla.ecommercee.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.List;
 
-@CrossOrigin("*")
 @RestController
 @RequestMapping("/v1/ecommercee/user/")
 public class UserController {
@@ -22,14 +23,14 @@ public class UserController {
         return userMapper.mapToUserDtoList(userService.getAllUsers());
     }
 
-    @PostMapping(value = "createUser")
-    public UserDto createUser(@RequestBody UserDto userDto) {
-        return userMapper.mapToUserDto(userService.save(userMapper.mapToUser(userDto)));
-    }
-
     @GetMapping(value = "getUser")
     public UserDto getUser(@RequestParam long id) throws UserNotFoundException {
         return userMapper.mapToUserDto(userService.getUser(id).orElseThrow(UserNotFoundException::new));
+    }
+
+    @PostMapping(value = "createUser")
+    public UserDto createUser(@RequestBody UserDto userDto) {
+        return userMapper.mapToUserDto(userService.save(userMapper.mapToUser(userDto)));
     }
 
     @PutMapping(value = "updateUser")
