+++ src/main/java/com/kodilla/ecommercee/controller/UserNotFoundException.java
@@ -1,5 +1,4 @@
 package com.kodilla.ecommercee.controller;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
