+++ src/main/java/com/kodilla/ecommercee/domain/Group.java
@@ -1,5 +1,4 @@
 package com.kodilla.ecommercee.domain;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
