+++ src/main/java/com/kodilla/ecommercee/domain/dto/CartDto.java
@@ -1,14 +1,12 @@
-package com.kodilla.ecommercee.dto;
-
+package com.kodilla.ecommercee.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import java.util.List;
 
-@Getter
 @AllArgsConstructor
 @NoArgsConstructor
+@Getter
 public class CartDto {
-
      private List<ProductDto> productList;
 }
