+++ src/main/java/com/kodilla/ecommercee/domain/dto/GroupDto.java
@@ -1,0 +1,12 @@
+package com.kodilla.ecommercee.domain.dto;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class GroupDto {
+    private Long id;
+    private String name;
+}
