+++ src/main/java/com/kodilla/ecommercee/domain/dto/OrderDto.java
@@ -1,9 +1,7 @@
-package com.kodilla.ecommercee.dto;
-
+package com.kodilla.ecommercee.domain.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 import java.time.LocalDate;
 import java.util.List;
 
@@ -11,7 +9,6 @@ import java.util.List;
 @NoArgsConstructor
 @Getter
 public class OrderDto {
-
     private Long id;
     private LocalDate date;
     private boolean isCompleted;
