--- src/main/java/com/kodilla/ecommercee/dto/ProductGroupDto.java
@@ -1,12 +1,0 @@
-package com.kodilla.ecommercee.dto;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ProductGroupDto {
-    private Long id;
-    private String name;
-}
