+++ src/main/java/com/kodilla/ecommercee/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee.mapper;
 import com.kodilla.ecommercee.domain.User;
-import com.kodilla.ecommercee.dto.UserDto;
+import com.kodilla.ecommercee.domain.dto.UserDto;
 import org.springframework.stereotype.Component;
 import java.util.List;
 import java.util.stream.Collectors;
