+++ src/main/java/com/kodilla/ecommercee/repository/GroupRepository.java
@@ -1,5 +1,4 @@
 package com.kodilla.ecommercee.repository;
-
 import com.kodilla.ecommercee.domain.Group;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
